+++ CHANGELOG.rst
@@ -13,6 +13,15 @@ General Information
 [2022.3.1] - 2023-02-17
 ***********************
 
+Changed
+=======
+- ``of_coloring`` now supports table group settings from ``of_multi_table``
+
+Added
+=====
+- Subscribed to new event ``kytos/of_multi_table.enable_table`` as well as publishing ``kytos/coloring.enable_table`` required to set a different ``table_id`` to flows.
+- Added ``settings.TABLE_GROUP_ALLOWED`` set containning the allowed table groups, for now there is only ``'base'``.
+
 Fixed
 =====
 - ``get_cookie`` could overflow 8 bytes for certain dpid values